restructure  to ui and action
